app/requirements/base.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.8
psycopg2-binary==2.8.6
pytz==2020.4
requests==2.22.0
sqlparse==0.4.1
urllib3==1.25.11